[chore] 파일 위치 변경 및 수정
