+++ Dibivava/Dibivava/Entites/TermDTO.swift
