+++ Dibivava/Dibivava/Network/SupplementNetworkService.swift
@@ -10,7 +10,7 @@ import Alamofire
 import RxSwift
 
 protocol SupplementNetworkService {
-    func requestSupplement(by id: Int) -> Single<SupplementDTO>
+    func requestSupplement(by id: Int) -> Single<SupplementResponse>
     func requestMaterial(by id: [String]?) -> Single<[MaterialDTO]?>
     func fetchTermDescription() -> Single<[TermDTO]>
 }
@@ -44,8 +44,8 @@ final class DefaultSupplementNetworkService: SupplementNetworkService {
         }
     }
     
-    func requestSupplement(by id: Int) -> Single<SupplementDTO> {
-        Single<SupplementDTO>.create { single in
+    func requestSupplement(by id: Int) -> Single<SupplementResponse> {
+        Single<SupplementResponse>.create { single in
             let urlString = "https://mp1878zrkj.execute-api.ap-northeast-2.amazonaws.com/dev/getSupplementById?id=\(id)"
             let urlComponent = URLComponents(string: urlString)
             guard let url = urlComponent?.url else { return Disposables.create() }
@@ -55,7 +55,7 @@ final class DefaultSupplementNetworkService: SupplementNetworkService {
                 case .success(let data):
                     do{
                         let decoder = JSONDecoder()
-                        let decodedData = try decoder.decode(SupplementDTO.self, from: data)
+                        let decodedData = try decoder.decode(SupplementResponse.self, from: data)
                         single(.success(decodedData))
                     }catch{
                         single(.failure(NetworkError.invalidStatusCode))
