+++ Dibivava/Dibivava/Presentation/HomeScene/View/HomeViewController.swift
@@ -17,7 +17,7 @@ import Kingfisher
 class HomeViewController: UIViewController {
     private let HomeViewmodel = HomeViewModel()
     private var searchresult: [Supplement] = []
-    private var searchInfo: [SupplementDTO] = []
+    private var searchInfo: [SupplementResponse] = []
     private let searchAPI = SearchAPI()
     private let warningLabel = UILabel().then{
         $0.text = "- [성분 정보 출저] 건강기능식품: 식품안전나라/ 건강기능식품 품목제조신고(원재료), 건강기능식품 기능성원료인정현황,  Ames, Bruce N; Gold, Lois Swirsky (2000).  “Paracelsus to parascience: The environmental cancer distraction”.  《Mutation Research/Fundamental and Molecular Mechanisms of Mutagenesis》 447: 3 - 본 정보는 참고용으로, 법적 책임을 지지 않습니다. - 본 정보는 참고용으로만 제공되며 개별적인 상황에 따라 반드시 의료전문가와 상담하여야합니다.어떠한 경우에도 본 앱의 내용을 근거로 한 자체 진단 또는                 치료를 시도해서는 안됩니다."
