+++ Dibivava/Dibivava/Presentation/SupplementDetailScene/View/BasePaddingLabel.swift
