+++ Dibivava/Dibivava/Presentation/SupplementDetailScene/View/MaterialType.swift
@@ -1,0 +1,14 @@
+//
+//  MaterialType.swift
+//  Dibivava
+//
+//  Created by dong eun shin on 2023/07/27.
+//
+
+import Foundation
+
+enum MaterialType: String {
+    case main
+    case sub
+    case addictive
+}
